Introduce static class properties
Since static class properties are tc39 stage 3 now https://github.com/tc39/proposal-class-fields and there are esdoc
and babel plugins for it, we could use it to improve the documentation for
static class properties and might think about declaring private class properties with
the hash tag. Both are things that are widely used in threeten.

Unfortunatly the most IDE's do not support the feature yet.