Fix LocalDateTime.of()
Calling LocalDateTime.ofNumbers with less then 3 number arguments resulted in an error,
because the defaults of month and dayOfMonth where set to 0, that leads to an exception.

threeten even requires at least 5 args here.

So now you have to either call of with a non null LocalDate and LocalTime or
with at least 3 valid numbers.