Fix bitwise or to logical compare to zero
This was a copy and paste bug from java to javascript, the bitwise or operation
do not work in javascript because its arguments are truncated to int32. This can lead to
ugly bugs for numbers greater then int32.

The issue is fixed by replacing with explicit comparision to zero.

Fixes #431