TemporalAccessor and Temporal are now explicit abstract base classes
TemporalAccessor got new "abstract" methods for documentation
purposes.

Temporal now has its implementation methods starting
with _ (underscore). The idea is to treat those methods as
"protected abstract", so derived classes are suppose to
implement them (or use the default implementation, if one
is available). Since they're now starting with _ is more
clearly expressed that are not meant to be called explicitly.