Remove generated distributions files from git
- Move all ci tests from travis to github actions
- Remove generated distributions in dist directory from git
- Adapt release howto
- Fix packaging, some packages contained unnecessary files

Motivation

Removing the generated distribution files from git, removes some complexity from the build and publish process,
because it's not required anymore to generate and commit distributions when bumping and tagging the version.
It's only required to change the version in the package.json and set a git tag, what is done by lerna version.
I hope, this pr is not breaking anything, but the packaged tars should look pretty much the same.

The PR is the next step to make the publish process more robust, less error prone and a preparation for moving it to ci.