Make it less easy to misuse the timezone module
I spent some time debugging

    ZoneId.of('Europe/Madrid')

crashing with

    DateTimeException: unsupported ZoneId:Europe/Madrid

this morning only to find out that this was actually because the default
getRules implementation always threw an exception like this.

I figured I'd try to improve things a bit by changing the error message
to be clearer (the problem is not related to any time zone id specifically
but it's connected to a missing import) and documenting things more
explicitly.